requirements.txt
fastapi
uvicorn
pymupdf
pdfminer.six
langchain
openai
faiss-cpu
sentence-transformers
streamlit
python-multipart
pydantic
tqdm
typing_extensions
